Introduce `ping_frequency` options
Allows `keep_alive` to be passed on to the server while `ping_frequency`
can optionally specify the frequency at which the actual ping messages
are sent. This follows Amazon's pattern of setting keep-alive to 20
minutes while pinging every 5 minutes:
https://aws.amazon.com/blogs/networking-and-content-delivery/implementing-long-running-tcp-connections-within-vpc-networking/

If `ping_frequency` is omitted, `keep_alive` will be used instead.